Move SourceGenerator to separate Helpers directory and add TestHelpers project
