+++ Testably.Abstractions.sln
@@ -6,9 +6,11 @@ MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_", "_", "{94F99274-3518-45D0-83A4-6F370D6FE74B}"
 	ProjectSection(SolutionItems) = preProject
 		.codacy.yml = .codacy.yml
+		.gitignore = .gitignore
 		CODE_OF_CONDUCT.md = CODE_OF_CONDUCT.md
 		CONTRIBUTING.md = CONTRIBUTING.md
 		Feature.Flags.props = Feature.Flags.props
+		global.json = global.json
 		README.md = README.md
 	EndProjectSection
 EndProject
@@ -72,7 +74,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Docs", "Docs", "{4D70CF83-2
 		Docs\Testing.md = Docs\Testing.md
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Testably.Abstractions.Tests.SourceGenerator", "Tests\Testably.Abstractions.Tests.SourceGenerator\Testably.Abstractions.Tests.SourceGenerator.csproj", "{97A2540A-7FB4-4D8B-B003-3F8CAA9CED87}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Helpers", "Helpers", "{B6C45D8A-A545-402E-A6B0-47BC7D9BBCF5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Testably.Abstractions.TestHelpers", "Tests\Helpers\Testably.Abstractions.TestHelpers\Testably.Abstractions.TestHelpers.csproj", "{425BA8BA-7795-4749-BFC1-6FA853EBC6B7}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Testably.Abstractions.Tests.SourceGenerator", "Tests\Helpers\Testably.Abstractions.Tests.SourceGenerator\Testably.Abstractions.Tests.SourceGenerator.csproj", "{97A2540A-7FB4-4D8B-B003-3F8CAA9CED87}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -120,6 +126,10 @@ Global
 		{5330AEEE-A915-4FF6-A85D-E9CCA161B654}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5330AEEE-A915-4FF6-A85D-E9CCA161B654}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5330AEEE-A915-4FF6-A85D-E9CCA161B654}.Release|Any CPU.Build.0 = Release|Any CPU
+		{425BA8BA-7795-4749-BFC1-6FA853EBC6B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{425BA8BA-7795-4749-BFC1-6FA853EBC6B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{425BA8BA-7795-4749-BFC1-6FA853EBC6B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{425BA8BA-7795-4749-BFC1-6FA853EBC6B7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{97A2540A-7FB4-4D8B-B003-3F8CAA9CED87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{97A2540A-7FB4-4D8B-B003-3F8CAA9CED87}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{97A2540A-7FB4-4D8B-B003-3F8CAA9CED87}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -140,7 +150,9 @@ Global
 		{1B078385-7AAA-4B8D-A88E-2CBCD3011F0D} = {61F335A6-9CE0-4040-A34F-E70B1A55077D}
 		{5330AEEE-A915-4FF6-A85D-E9CCA161B654} = {61F335A6-9CE0-4040-A34F-E70B1A55077D}
 		{4D70CF83-2EEE-439E-AA30-3FADAACAEB0C} = {94F99274-3518-45D0-83A4-6F370D6FE74B}
-		{97A2540A-7FB4-4D8B-B003-3F8CAA9CED87} = {61F335A6-9CE0-4040-A34F-E70B1A55077D}
+		{B6C45D8A-A545-402E-A6B0-47BC7D9BBCF5} = {61F335A6-9CE0-4040-A34F-E70B1A55077D}
+		{425BA8BA-7795-4749-BFC1-6FA853EBC6B7} = {B6C45D8A-A545-402E-A6B0-47BC7D9BBCF5}
+		{97A2540A-7FB4-4D8B-B003-3F8CAA9CED87} = {B6C45D8A-A545-402E-A6B0-47BC7D9BBCF5}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {EC4D8481-B9FD-41B5-832A-369210993DF4}
