+++ Tests/Helpers/Testably.Abstractions.Tests.SourceGenerator/ClassGeneratorBase.cs
