+++ Tests/Helpers/Testably.Abstractions.Tests.SourceGenerator/ClassGenerators/RandomSystemClassGenerator.cs
@@ -3,24 +3,24 @@ using Testably.Abstractions.Tests.SourceGenerator.Model;
 
 namespace Testably.Abstractions.Tests.SourceGenerator.ClassGenerators;
 
-internal class TimeSystemClassGenerator : ClassGeneratorBase
+internal class RandomSystemClassGenerator : ClassGeneratorBase
 {
 	/// <inheritdoc cref="ClassGeneratorBase.Marker" />
 	public override string Marker
-		=> "TimeSystemTestBase<TTimeSystem>";
+		=> "RandomSystemTestBase<TRandomSystem>";
 
 	/// <inheritdoc cref="ClassGeneratorBase.GenerateSource(StringBuilder, ClassModel)" />
 	protected override void GenerateSource(StringBuilder sourceBuilder, ClassModel @class)
 		=> sourceBuilder.Append(@$"
-using Testably.Abstractions.Tests.TestHelpers;
+using Testably.Abstractions.TestHelpers;
 using Xunit.Abstractions;
 
 namespace {@class.Namespace}
 {{
-	public abstract partial class {@class.Name}<TTimeSystem>
+	public abstract partial class {@class.Name}<TRandomSystem>
 	{{
-		protected {@class.Name}(TTimeSystem timeSystem)
-			: base(timeSystem)
+		protected {@class.Name}(TRandomSystem randomSystem)
+			: base(randomSystem)
 		{{
 		}}
 	}}
@@ -29,17 +29,17 @@ namespace {@class.Namespace}
 namespace {@class.Namespace}.{@class.Name}
 {{
 	// ReSharper disable once UnusedMember.Global
-	public sealed class MockTimeSystemTests : {@class.Name}<MockTimeSystem>
+	public sealed class MockRandomSystemTests : {@class.Name}<MockRandomSystem>
 	{{
-		public MockTimeSystemTests() : base(new MockTimeSystem(TimeProvider.Now()))
+		public MockRandomSystemTests() : base(new MockRandomSystem())
 		{{
 		}}
 	}}
 
 	// ReSharper disable once UnusedMember.Global
-	public sealed class RealTimeSystemTests : {@class.Name}<RealTimeSystem>
+	public sealed class RealRandomSystemTests : {@class.Name}<RealRandomSystem>
 	{{
-		public RealTimeSystemTests() : base(new RealTimeSystem())
+		public RealRandomSystemTests() : base(new RealRandomSystem())
 		{{
 		}}
 	}}
