+++ Tests/Helpers/Testably.Abstractions.Tests.SourceGenerator/Model/ClassModel.cs
